include rhs changes to test solvers
