+++ funtofem/interface/test_solver.py
@@ -377,25 +377,28 @@ class TestAerodynamicSolver(SolverInterface):
             aero_disps_ajp = body.get_aero_disps_ajp(scenario)
             if aero_loads_ajp is not None:
                 for k, func in enumerate(scenario.functions):
+                    if include_rhs and func.analysis_type == "aerodynamic":
+                        aero_loads_ajp[:, k] = self.scenario_data[
+                            scenario.id
+                        ].func_coefs1
                     aero_disps_ajp[:, k] = np.dot(
                         self.scenario_data[scenario.id].Jac1.T, aero_loads_ajp[:, k]
                     )
-                    if include_rhs and func.analysis_type == "aerodynamic":
-                        aero_disps_ajp[:, k] += self.scenario_data[
-                            scenario.id
-                        ].func_coefs1
 
             aero_flux_ajp = body.get_aero_heat_flux_ajp(scenario)
             aero_temps_ajp = body.get_aero_temps_ajp(scenario)
             if aero_flux_ajp is not None:
                 for k, func in enumerate(scenario.functions):
+                    if include_rhs and func.analysis_type == "aerodynamic":
+                        aero_flux_ajp[:, k] = self.scenario_data[
+                            scenario.id
+                        ].func_coefs2
                     aero_temps_ajp[:, k] = np.dot(
                         self.scenario_data[scenario.id].Jac2.T, aero_flux_ajp[:, k]
                     )
-                    if include_rhs and func.analysis_type == "aerodynamic":
-                        aero_temps_ajp[:, k] += self.scenario_data[
-                            scenario.id
-                        ].func_coefs2
+
+        if step != 0:
+            self.get_function_gradients(scenario, bodies)
 
         fail = 0
         return fail
@@ -589,6 +592,20 @@ class TestStructuralSolver(SolverInterface):
                         )
                         func.add_gradient_component(var, value)
 
+                # debug
+                debug = False
+                if debug:
+                    deriv = func.get_gradient_component(var)
+                    struct_loads_ajp = body.get_struct_loads_ajp(scenario)
+                    print(f"struct loads ajp deriv = {struct_loads_ajp[:3]}")
+                    print(
+                        f"struct loads ajp norm = {np.linalg.norm(struct_loads_ajp):.4f}"
+                    )
+                    print(
+                        f"struct disps ajp norm = {np.linalg.norm(struct_disps_ajp):.4f}"
+                    )
+                    print(f"df{findex}/dv{vindex} = {deriv}", flush=True)
+
         return
 
     def get_coordinate_derivatives(self, scenario, bodies, step):
@@ -707,25 +724,33 @@ class TestStructuralSolver(SolverInterface):
             struct_loads_ajp = body.get_struct_loads_ajp(scenario)
             if struct_disps_ajp is not None:
                 for k, func in enumerate(scenario.functions):
+                    if include_rhs and func.analysis_type == "structural":
+                        struct_disps_ajp[:, k] = self.scenario_data[
+                            scenario.id
+                        ].func_coefs1
                     struct_loads_ajp[:, k] = np.dot(
                         self.scenario_data[scenario.id].Jac1.T, struct_disps_ajp[:, k]
                     )
-                    if include_rhs and func.analysis_type == "structural":
-                        struct_loads_ajp[:, k] += self.scenario_data[
-                            scenario.id
-                        ].func_coefs1
+
+                # print(f"struct disps ajp iter_adj = {struct_disps_ajp[:3]}", flush=True)
+                # print(f"struct loads ajp iter_adj = {struct_loads_ajp[:3]}", flush=True)
 
             struct_temps_ajp = body.get_struct_temps_ajp(scenario)
             struct_flux_ajp = body.get_struct_heat_flux_ajp(scenario)
             if struct_temps_ajp is not None:
                 for k, func in enumerate(scenario.functions):
+                    if include_rhs and func.analysis_type == "structural":
+                        struct_temps_ajp[:, k] = self.scenario_data[
+                            scenario.id
+                        ].func_coefs2
                     struct_flux_ajp[:, k] = np.dot(
                         self.scenario_data[scenario.id].Jac2.T, struct_temps_ajp[:, k]
                     )
-                    if include_rhs and func.analysis_type == "structural":
-                        struct_flux_ajp[:, k] += self.scenario_data[
-                            scenario.id
-                        ].func_coefs2
+                # print(f"struct flux ajp = {struct_disps_ajp}")
+
+        # add derivative values
+        if step != 0:
+            self.get_function_gradients(scenario, bodies)
 
         fail = 0
         return fail
